# ApiDockerTecnimotors/ApiDockerTecnimotors/obj/Debug/net8.0/EndpointInfo/ApiDockerTecnimotors.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ApiDockerTecnimotors",
    "version": "1.0"
  },
  "paths": {
...
    "/api/WishList/RegisterWishList": {
      "post": {
        "tags": [
          "WishList"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrModels"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TrModels"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TrModels"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/WishList/ListadoWishList": {
      "post": {
        "tags": [
          "WishList"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrUuid"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TrUuid"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TrUuid"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/WishList/NextOrderNumber": {
      "post": {
        "tags": [
          "WishList"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Trtrial"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Trtrial"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Trtrial"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/WishList/removeWishList": {
      "post": {
        "tags": [
          "WishList"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveFromWishlistRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveFromWishlistRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveFromWishlistRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/WishList/RegistrarWishlogin": {
      "post": {
        "tags": [
          "WishList"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListWishRegister"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ListWishRegister"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ListWishRegister"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ListWish": {
        "type": "object",
        "properties": {
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "unidad": {
            "type": "string",
            "nullable": true
          },
          "categoria": {
            "type": "string",
            "nullable": true
          },
          "marca": {
            "type": "string",
            "nullable": true
          },
          "marcaoriginal": {
            "type": "string",
            "nullable": true
          },
          "medida": {
            "type": "string",
            "nullable": true
          },
          "modelo": {
            "type": "string",
            "nullable": true
          },
          "medidaestandarizado": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "codigo": {
            "type": "string",
            "nullable": true
          },
          "familia": {
            "type": "string",
            "nullable": true
          },
          "subfamilia": {
            "type": "string",
            "nullable": true
          },
          "tipo": {
            "type": "string",
            "nullable": true
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "producto": {
            "type": "string",
            "nullable": true
          },
          "vendor": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "pathimagen": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ListWishRegister": {
        "type": "object",
        "properties": {
          "listWishLogin": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListWish"
            },
            "nullable": true
          },
          "uuidcliente": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RemoveFromWishlistRequest": {
        "type": "object",
        "properties": {
          "uuidcliente": {
            "type": "string",
            "nullable": true
          },
          "codigo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TrModels": {
        "type": "object",
        "properties": {
          "idwishlish": {
            "type": "integer",
            "format": "int32"
          },
          "uuidcliente": {
            "type": "string",
            "nullable": true
          },
          "ordenwishlist": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "unidad": {
            "type": "string",
            "nullable": true
          },
          "categoria": {
            "type": "string",
            "nullable": true
          },
          "marca": {
            "type": "string",
            "nullable": true
          },
          "marcaoriginal": {
            "type": "string",
            "nullable": true
          },
          "medida": {
            "type": "string",
            "nullable": true
          },
          "modelo": {
            "type": "string",
            "nullable": true
          },
          "medidaestandarizado": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "codigo": {
            "type": "string",
            "nullable": true
          },
          "familia": {
            "type": "string",
            "nullable": true
          },
          "subfamilia": {
            "type": "string",
            "nullable": true
          },
          "tipo": {
            "type": "string",
            "nullable": true
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "producto": {
            "type": "string",
            "nullable": true
          },
          "vendor": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "pathimagen": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "type": "string",
            "nullable": true
          },
          "fecharegistro": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TrUuid": {
        "type": "object",
        "properties": {
          "uuidcliente": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Trtrial": {
        "type": "object",
        "properties": {
          "uuidcliente": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}